requirements.txt
# ---- Core Web Framework ----
fastapi==0.127.0
uvicorn[standard]==0.30.6

# ---- Database (SQLite + SQLAlchemy) ----
sqlalchemy==2.0.32

# ---- Background Jobs ----
celery==5.6.0
redis==7.1.0

# ---- OpenTelemetry Core ----
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0

# ---- OpenTelemetry Instrumentation ----
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-celery==0.46b0
opentelemetry-instrumentation-redis==0.46b0

# ---- Config & Validation ----
pydantic==2.9.2
pydantic-settings==2.5.2
